refactor: Move linker to utilities. Remove excess where possible
